Most Menu Operations Work As Follows:
Main Menu:
Use the mouse to navigate and choose option (needs to be fixed)

Instructions:
Use the A button to switch instruction pages, Use the B button to go back to the main munu:

Credits:
Use the B button to go back to the main menu.

InGame:
Use the start button to open the puase menu, where the player can choose to either resume, or go back to the title screen using the mouse (needs to be fixed.)

NOTE: The 'instructions' option in the pause menu is not functioning yet.